environment control
we used python decouple for environment control. installed it on the docker container using command below in terminal:
docker-compose exec backend sh -c "pip install python-decouple"
docker container must be up. added it to the requirements.txt of  the project.
then to use it we change the setting file of docker container in the core\core folder.
modified environment section in the docker compose to set the feed values for the sections we set in the settings.py above.